+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 # [Unrealeased]
+- [Collection read regex name](https://github.com/EruEri/oyomu/pull/10)
 - [Improve Kitty, iTerm render](https://github.com/EruEri/oyomu/pull/9)
 
 # [0.2.2-1]